Ingestion/requirements.txt
azure-storage-blob==12.26.0